+++ src/relstorage/storage/copy.py
@@ -45,10 +45,17 @@ class Copy(object):
         self.tpc = tpc
         self.restore = restore
 
+    # Time in seconds between progress logging.
+    log_interval = 60
+
+    # Number of transactions to copy before checking if we should log.
+    log_count = 10
+
     def copyTransactionsFrom(self, other):
-        # pylint:disable=too-many-locals,too-many-statements
+        # pylint:disable=too-many-locals,too-many-statements,too-many-branches
         # adapted from ZODB.blob.BlobStorageMixin
         begin_time = time.time()
+        log_at = begin_time + self.log_interval
         txnum = 0
         total_size = 0
         blobhelper = self.blobhelper
@@ -57,6 +64,7 @@ class Copy(object):
 
         logger.info("Counting the transactions to copy.")
         other_it = other.iterator()
+        logger.debug("Opened the other iterator: %s", other_it)
         try:
             num_txns = len(other_it)
             if num_txns == 0:
@@ -64,6 +72,7 @@ class Copy(object):
                 # Try the other path.
                 raise TypeError()
         except TypeError:
+            logger.debug("Iterator %s doesn't support len()", other_it)
             num_txns = 0
             for _ in other_it:
                 num_txns += 1
@@ -71,6 +80,7 @@ class Copy(object):
             other_it = other.iterator()
         logger.info("Copying %d transactions", num_txns)
 
+        tmp_blobs_to_rm = []
         for trans in other_it:
             txnum += 1
             num_txn_records = 0
@@ -87,9 +97,13 @@ class Copy(object):
                 if blobfile is not None:
                     fd, name = tempfile.mkstemp(
                         suffix='.tmp',
-                        dir=blobhelper.temporaryDirectory())
-                    os.close(fd)
-                    with open(name, 'wb') as target:
+                        dir=blobhelper.temporaryDirectory()
+                    )
+                    tmp_blobs_to_rm.append(name)
+                    logger.debug("Copying %s to temporary blob file %s for upload",
+                                 blobfile, name)
+
+                    with os.fdopen(fd, 'wb') as target:
                         copy_blob(blobfile, target)
                     blobfile.close()
                     restore.restoreBlob(record.oid, record.tid, record.data,
@@ -103,17 +117,30 @@ class Copy(object):
             tpc.tpc_vote(trans)
             tpc.tpc_finish(trans)
 
-            pct_complete = '%1.2f%%' % (txnum * 100.0 / num_txns)
-            elapsed = time.time() - begin_time
-            if elapsed:
-                rate = total_size / 1e6 / elapsed
-            else:
-                rate = 0.0
-            rate_str = '%1.3f' % rate
-            logger.info(
-                "Copied tid %d,%5d records | %6s MB/s (%6d/%6d,%7s)",
-                bytes8_to_int64(trans.tid), num_txn_records, rate_str,
-                txnum, num_txns, pct_complete)
+            for tmp_blob in tmp_blobs_to_rm:
+                logger.debug("Removing temporary blob file %s", tmp_blob)
+                try:
+                    os.unlink(tmp_blob)
+                except OSError:
+                    pass
+            del tmp_blobs_to_rm[:]
+
+            if txnum % self.log_count == 0 and time.time() > log_at:
+                now = time.time()
+                log_at = now + self.log_interval
+
+                pct_complete = '%1.2f%%' % (txnum * 100.0 / num_txns)
+                elapsed = now - begin_time
+                if elapsed:
+                    rate = total_size / 1e6 / elapsed
+                else:
+                    rate = 0.0
+                rate_str = '%1.3f' % rate
+
+                logger.info(
+                    "Copied tid %d,%5d records | %6s MB/s (%6d/%6d,%7s)",
+                    bytes8_to_int64(trans.tid), num_txn_records, rate_str,
+                    txnum, num_txns, pct_complete)
 
         elapsed = time.time() - begin_time
         logger.info(